@@ -12,7 +12,7 @@ func TestFolderDelete(t *testing.T) {
 
 	tests := []struct {
 		give         string
-		wantReadBack map[string]map[string]interface{}
+		wantReadBack map[string]map[string]any
 		wantErr      []error
 	}{
 		{
@@ -32,7 +32,7 @@ func TestFolderDelete(t *testing.T) {
 		},
 		{
 			give: "0/4/13/24/25/error/list/inject",
-			wantReadBack: map[string]map[string]interface{}{
+			wantReadBack: map[string]map[string]any{
 				"26/27": {
 					"28": "29",
 				},
@@ -41,7 +41,7 @@ func TestFolderDelete(t *testing.T) {
 		},
 		{
 			give: "0/4/13/24/25/26/error/delete/inject",
-			wantReadBack: map[string]map[string]interface{}{
+			wantReadBack: map[string]map[string]any{
 				"27": {
 					"28": "29",
 				},
