@@ -2,5 +2,5 @@ package vaku
 
 // Version gives the current Vaku API version.
 func Version() string {
-	return "2.4.5"
+	return "2.4.6"
 }
