@@ -8,5 +8,5 @@ import (
 
 func TestVersion(t *testing.T) {
 	t.Parallel()
-	assert.Equal(t, "2.4.5", Version())
+	assert.Equal(t, "2.4.6", Version())
 }
